Feature diameter in GAL file is double
git-svn-id: https://subversion.biologie.ens.fr/svn/leburon/trunk/nividic@281 5a62d92b-e016-0410-9103-b89fe532bb34

+++ src/main/java/fr/ens/transcriptome/nividic/om/ArrayBlock.java
@@ -143,7 +143,7 @@ public class ArrayBlock {
    * Set the diameter of the features of the block.
    * @param featureDiameter The featureDiameter to set
    */
-  public void setFeatureDiameter(final int featureDiameter) {
+  public void setFeatureDiameter(final double featureDiameter) {
     this.featureDiameter = featureDiameter;
   }
